data-scheduler/pyproject.toml
[project]
name = "data-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv[cli]>=0.9.9",
    "requests>=2.32.5",
    "rich>=14.2.0",
]
